pyproject.toml
[project]
name = "lynx-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.5",
    "discord-py>=2.7.1",
    "python-dotenv>=1.2.2",
]
